# booking api/tests/src/main/resources/api-spec.yaml
openapi: 3.0.1
info:
  title: Bookings API
  description: API to book activities at different locations.
  version: 1.0.0
servers:
  - url: https://192.168.99.100/api
...
paths:
...
  /location/all:
    get:
      tags:
        - locations
      summary: Get all locations
      operationId: getLocations
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          description: page wished, defaults to first page (0).
          required: false
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Location'
        406:
          description: Bad accept header
  /location:
    post:
      tags:
        - locations
      summary: Add a new location
      operationId: addLocation
      requestBody:
        description: New location to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Location'
        required: true
      responses:
        204:
          description: Location created successfully
        400:
          description: Invalid input
        401:
          description: Auth token required
        403:
          description: Forbidden
      security:
        - bearerAuth: []
  /location/{id}:
    put:
      tags:
        - locations
      summary: Update a location, id must exist
      operationId: updateLocation
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: id of location to edit.
          required: true
      requestBody:
        description: New values for location
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Location'
        required: true
      responses:
        204:
          description: Updated successfully
        400:
          description: Invalid input
        401:
          description: Auth token required
        403:
          description: Forbidden
        404:
          description: id not found
      security:
        - bearerAuth: []
    delete:
      tags:
        - locations
      summary: Delete the specified location, id must exist
      operationId: deleteLocation
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: id of location to delete
          required: true
      responses:
        204:
          description: Deleted successfully
        401:
          description: Auth token required
        403:
          description: Forbidden
        404:
          description: id not found
      security:
        - bearerAuth: []

      ##### bookings #####
  /booking/my:
    get:
      tags:
        - bookings
      summary: get logged user's bookings
      operationId: getBookings
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          description: page wished, defaults to first page (0).
          required: false
        - in: query
          name: previous
          schema:
            type: boolean
          description: true to get previous bookings.
          required: false
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
        401:
          description: Auth token required
        406:
          description: Bad accept header
      security:
        - bearerAuth: []
...
  /booking/{id}:
    delete:
      tags:
        - bookings
      summary: deletes booking for logged user
      operationId: deleteBooking
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: id of booking to delete
          required: true
      responses:
        204:
          description: successfully deleted
        401:
          description: must be logged
        403:
          description: unauthorized
        404:
          description: booking not found
      security:
        - bearerAuth: []
components:
  schemas:
...
    Location:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          minLength: 3
        address:
          type: string
          minLength: 15
      required:
        - name
        - address
    Booking:
      type: object
      properties:
        id:
          type: integer
        activityId:
          type: integer
        activityName:
          type: string
        locationId:
          type: integer
        locationName:
          type: string
        date:
          type: string
        nbPlaces:
          type: integer
      required:
        - activityId
        - locationId
        - date
        - nbPlaces
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT 